+++ CHANGELOG.md
@@ -5,6 +5,76 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.0.5] - 15-09-2022
+
+### Changed
+
+- `nova_get_menus` now eagerloads menu items.
+
+## [7.0.14] - 13-09-2022
+
+### Changed
+
+- Misc UI fixes and improvements
+
+## [7.0.13] - 12-09-2022
+
+### Changed
+
+- Misc UI fixes and improvements
+- Updated packages
+
+## [7.0.12] - 31-08-2022
+
+### Added
+
+- Added `title`, `icon`, `hideMenu` functions to `MenuBuilder` class to help customize sidebar menu.
+- Added `nova_get_menu_by_id` helper. (thanks to [@ttungbmt](https://github.com/ttungbmt))
+
+### Changed
+
+- Removed `overflow-hidden` from UpdateModel. Fixes an issue where dropdown fields were cut-off.
+
+## [7.0.11] - 29-08-2022
+
+### Changed
+
+- Fixed a rare UI issue with menu builder field width
+
+## [7.0.10] - 09-08-2022
+
+### Changed
+
+- Fixed an issue where static-url type menu item did not display `URL` field.
+
+## [7.0.9] - 27-07-2022
+
+### Changed
+
+- Move "open in tab/window" field before custom fields
+
+## [7.0.8] - 27-07-2022
+
+### Added
+
+- Added value emit event to "name" field
+
+### Changed
+
+- Fixed some validation errors
+- Made fields stacked for a cleaner look
+- Updated packages
+
+## [7.0.7] - 21-07-2022
+
+### Added
+
+- Added support for `->rules()`, `->updateRules()`, `->creationRules()` for fields in `getFields()`
+
+### Changed
+
+- Updated packages
+
 ## [7.0.6] - 17-05-2022
 
 ### Changed